fix(build): remove dacc run from build, by moving it inside lib folder
