+++ src/drive/targets/services/dacc.js
@@ -1,6 +1,6 @@
 import log from 'cozy-logger'
 import fetch from 'node-fetch'
-import { run } from './dacc-run'
+import { run } from 'lib/dacc/dacc-run'
 
 global.fetch = fetch
